[Move & Remove] 파일 변경 및 삭제 #17
